feat: move Azurite container to the new Azure module, deprecating the old one
