+++ modules/azure/Makefile
@@ -1,0 +1,5 @@
+include ../../commons-test.mk
+
+.PHONY: test
+test:
+	$(MAKE) test-azure
