+++ modules/azure/azurite/azurite_test.go
@@ -1,0 +1,21 @@
+package azurite_test
+
+import (
+	"context"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/testcontainers/testcontainers-go"
+	"github.com/testcontainers/testcontainers-go/modules/azurite"
+)
+
+func TestAzurite(t *testing.T) {
+	ctx := context.Background()
+
+	ctr, err := azurite.Run(ctx, "mcr.microsoft.com/azure-storage/azurite:3.23.0")
+	testcontainers.CleanupContainer(t, ctr)
+	require.NoError(t, err)
+
+	// perform assertions
+}
