+++ modules/azure/azurite/services.go
@@ -1,0 +1,9 @@
+package azurite
+
+const (
+	BlobService  Service = "blob"
+	QueueService Service = "queue"
+	TableService Service = "table"
+)
+
+type Service string
