+++ modules/azurite/options.go
@@ -1,34 +1,14 @@
 package azurite
 
 import (
-	"fmt"
-
 	"github.com/testcontainers/testcontainers-go"
+	"github.com/testcontainers/testcontainers-go/modules/azure/azurite"
 )
 
-type options struct {
-	// EnabledServices is a list of services that should be enabled
-	EnabledServices []Service
-}
-
-func defaultOptions() options {
-	return options{
-		EnabledServices: []Service{BlobService, QueueService, TableService},
-	}
-}
-
+// Deprecated: This option is deprecated in favor of the one in "modules/azure/azurite".
+// Please use that package instead for all new code.
 // WithInMemoryPersistence is a custom option to enable in-memory persistence for Azurite.
 // This option is only available for Azurite v3.28.0 and later.
 func WithInMemoryPersistence(megabytes float64) testcontainers.CustomizeRequestOption {
-	return func(req *testcontainers.GenericContainerRequest) error {
-		cmd := []string{"--inMemoryPersistence"}
-
-		if megabytes > 0 {
-			cmd = append(cmd, "--extentMemoryLimit", fmt.Sprintf("%f", megabytes))
-		}
-
-		req.Cmd = append(req.Cmd, cmd...)
-
-		return nil
-	}
+	return azurite.WithInMemoryPersistence(megabytes)
 }
